Using Pydantic TraceSpan model
