+++ src/judgeval/common/tracer.py
@@ -88,116 +88,6 @@ class EvaluationConfig:
     log_results: Optional[bool] = True
 # --- End Evaluation Config Dataclass ---
 
-@dataclass
-class TraceSpan:
-    """Represents a single trace span."""
-    span_id: str
-    trace_id: str
-    depth: int
-    created_at: float
-    inputs: dict = field(default_factory=dict)
-    output: Any = None
-    evaluation_runs: List[EvaluationRun] = field(default_factory=list)
-    parent_span_id: Optional[str] = None
-    function: Optional[str] = None
-    duration: Optional[float] = None
-    span_type: SpanType = "span"
-    message: Optional[str] = None
-
-    def print_span(self):
-        """Print the span with proper formatting and parent relationship information."""
-        indent = "  " * self.depth
-        parent_info = f" (parent_id: {self.parent_span_id})" if self.parent_span_id else ""
-        print(f"{indent}→ {self.function} (id: {self.span_id}){parent_info} (trace: {self.message})")
-    
-    def to_dict(self) -> dict:
-        """Convert the trace span to a dictionary format for storage/transmission."""
-        return {
-            "span_id": self.span_id,
-            "trace_id": self.trace_id,
-            "depth": self.depth,
-            "created_at": datetime.fromtimestamp(self.created_at).isoformat(),
-            "inputs": self._serialize_inputs(),
-            "output": self._serialize_output(),
-            "evaluation_runs": [run.model_dump() for run in self.evaluation_runs] if self.evaluation_runs else [],
-            "parent_span_id": self.parent_span_id,
-            "function": self.function,
-            "duration": self.duration,
-            "span_type": self.span_type,
-            "message": self.message
-        }
-    
-    def _serialize_inputs(self) -> dict:
-        """Helper method to serialize input data safely."""
-        serialized_inputs = {}
-        for key, value in self.inputs.items():
-            if isinstance(value, BaseModel):
-                serialized_inputs[key] = value.model_dump()
-            elif isinstance(value, (list, tuple)):
-                # Handle lists/tuples of arguments
-                serialized_inputs[key] = [
-                    item.model_dump() if isinstance(item, BaseModel)
-                    else None if not self._is_json_serializable(item)
-                    else item
-                    for item in value
-                ]
-            else:
-                if self._is_json_serializable(value):
-                    serialized_inputs[key] = value
-                else:
-                    serialized_inputs[key] = self.safe_stringify(value, self.function)
-        return serialized_inputs
-
-    def _is_json_serializable(self, obj: Any) -> bool:
-        """Helper method to check if an object is JSON serializable."""
-        try:
-            json.dumps(obj)
-            return True
-        except (TypeError, OverflowError, ValueError):
-            return False
-
-    def safe_stringify(self, output, function_name):
-        """
-        Safely converts an object to a string or repr, handling serialization issues gracefully.
-        """
-        try:
-            return str(output)
-        except (TypeError, OverflowError, ValueError):
-            pass
-    
-        try:
-            return repr(output)
-        except (TypeError, OverflowError, ValueError):
-            pass
-    
-        warnings.warn(
-            f"Output for function {function_name} is not JSON serializable and could not be converted to string. Setting to None."
-        )
-        return None
-        
-    def _serialize_output(self) -> Any:
-        """Helper method to serialize output data safely."""
-        def serialize_value(value):
-            if isinstance(value, BaseModel):
-                return value.model_dump()
-            elif isinstance(value, dict):
-                # Recursively serialize dictionary values
-                return {k: serialize_value(v) for k, v in value.items()}
-            elif isinstance(value, (list, tuple)):
-                # Recursively serialize list/tuple items
-                return [serialize_value(item) for item in value]
-            else:
-                # Try direct JSON serialization first
-                try:
-                    json.dumps(value)
-                    return value
-                except (TypeError, OverflowError, ValueError):
-                    # Fallback to safe stringification
-                    return self.safe_stringify(value, self.function)
-
-        # Start serialization with the top-level output
-        return serialize_value(self.output)
-
 # Temporary as a POC to have log use the existing annotations feature until log endpoints are ready
 @dataclass
 class TraceAnnotation:
@@ -765,7 +655,7 @@ class TraceClient:
             "project_name": self.project_name,
             "created_at": datetime.utcfromtimestamp(self.start_time).isoformat(),
             "duration": total_duration,
-            "entries": [span.to_dict() for span in self.trace_spans],
+            "entries": [span.model_dump() for span in self.trace_spans],
             "evaluation_runs": [run.model_dump() for run in self.evaluation_runs],
             "overwrite": overwrite,
             "parent_trace_id": self.parent_trace_id,
